Introduce nitrokey-test crate
Running tests against Nitrokey devices is a cumbersome process because
of all the preconditions that have to be checked:
- some tests work only on a specific device (e.g., the Nitrokey
  Storage), while others test generic functionality
- if both device types are supported, the test should likely run on both
  to ensure proper coverage (after all, the firmware is different on both
  devices)
- a developer may have only one type of device available, a condition
  that can be detected most reliably at runtime and without the need for
  user interaction

This change introduces a crate that takes care of the automated handling
of those cases. It provides a declarative interface for a developer to
specify what device a test can run on and will ensure to that it is run
on such a device, if present.