Skip tests when a viable Nitrokey device is not present
Ultimately, detecting devices at runtime, while already providing some
advantages, really only plays out all its benefits if a device that is
currently not present does not cause a test failure. Instead, the test
should be skipped. Unfortunately, Rust does not have the concept of
skipped tests. It does support ignoring them, but that is a compile time
decision which just gets us back to square one.
In order to still be usable, this change makes tests succeed if a device
is found to be not present. To make it a bit more obvious, we also print
a string indicating that the test did not actually run, but was in fact
skipped.