Add support for tests that work only if no device is present
There is a small number of tests that make only sense to be run when no
device is present, because they test error behavior when no device can
be found.
In order to support those, this change adds code to recognize a new test
function signature: one without arguments. Such a test function is only
run when no device is present.