Introduce the concept of a device group
Another requirement we have is to support running of tests only for a
particular device upon request. That is, if the user desires to only run
tests for the Nitrokey Pro, we should honor this request and not even
attempt to run tests that target the Nitrokey Storage.
This change does not yet introduce the full logic required to
accommodate such requests, but it adds a piece that is required.
Specifically, we introduce the concept of a device group. There are
three groups, encompassing the tests when a Nitrokey Pro, Nitrokey
Storage, or no device at all are present.