Honor NITROKEY_TEST_GROUP environment variable
Despite aiming for ease of usability and running all tests by default
(dependent on device availability, of course), we would like to provide
the user with a mechanism to only run tests of a certain group. We
already introduced the concept of a device group and with this change we
honor the NITROKEY_TEST_GROUP variable. This variable can either be unset (the
existing behavior is preserved) or represents