Rename test_device macro to test
While test_device is a reasonable choice of name for the macro this
crate provides, it is opinionated in an unnecessary way. As in recent
versions of Rust macros can now be imported directly from the providing
crate, it seems preferable to just name the macro of interest "test",
and allow clients to rename as per the usual mechanism.