Properly use absolutely qualified macros
The code generated by the macro runs in the context of the client using
the crate. We should make every effort to reference the proper standard
symbols we intend to use in an absolute manner, or we run risk of
picking up customizations that the client code made.
In particular, we have seen a case where client code replaced the
println!() macro with a custom definition accepting an additional
parameter. By merely referencing the macro without full qualification,
we picked up this client-side change unintentionally.