Use dedicated crate to provide mutual exclusion related state
In order to properly serialize tests what we have done so far was to
emit a function into the crate under test that provides access to a
shared (static) mutex.
Unfortunately, doing so does not work very well when multiple test
modules want to use nitrokey-test. The problem is that the function
being emitted is local to one of those modules and there is no way to
reference it from other modules without a-priori knowledge about the
module structure.
In an attempt to sidestep this problem, this change replaces the
function being emitted with a dedicated crate providing the necessary
state to achieve mutual exclusion for us. Because the crate has a
globally unique name, it can be referenced just fine, irrespective of
what module does so (first).