Properly handle mut modifier in test function signature
Right now we use a fairly rigid approach for invoking the actual test
function the user specified: we simply create a variable with the same
name the user's test function signature has. Part of this rigidness
manifests in mutability not being handled properly. That is, we silently
discard any mut qualifiers in the test function signature.
With this change we fix this problem by explicitly checking for
mutability and declaring our variable as mutable accordingly.