Emit dedicated function for test function body
So far we have treated the user's test function body as a white box,
assuming that it expects a variable called 'device' representing the
Nitrokey device. While that is not very limiting in practice, it still
is an unnecessary restriction and even leads to additional complexity in
the code in this crate.
To fix that, this change instead emits a function that contains the test
function's body and just invokes this function with the appropriate
device. This approach has the added advantage that we no longer need the
explicit handling of mutable vs. non-mutable arguments to the test
function.