Add integration style tests for basic test function invocation
This change adds tests for the core functionality that the crate
provides: the connection to a connected Nitrokey along with the
invocation of the appropriate test.
Unfortunately we are limited by physical factors here in achieving full
test coverage. In particular, we can only verify that if a test function
is invoked that is has been passed the right data. We can't do much
about a function not being invoked at all, because while we can panic
from a test and expect a panic from it, we ultimately need a physical
device present (or not present, depending on the test) to trigger the
case.