Streamline test function call code
This change reworks the expand_call function a bit, streamlining the
connection logic contained in it. We do so by introducing logic to map
from an EmittedDevice to a DeviceGroup and invoking expand_connect based
on the device group inferred that way.
We also separate the connection code from the test invocation. This
separation uncovered a bug due to which we did not actually test the
function in the no device case. Unfortunately no way was found to test
this part in an automated fashion, because it fundamentally relies on
physical constraints, i.e., no Nitrokey being present.